+++ src/backend/resolvers/mutations/sitemap/permissions-upsertion/tests/upsertUserPerms/mocks/input/regimens.js
@@ -1,0 +1,16 @@
+const { ObjectId } = require('mongodb')
+
+module.exports = [
+  {
+    _id: ObjectId('1628394965aa'),
+    name: 'Integration test regimen 1',
+  },
+  {
+    _id: ObjectId('1628394966aa'),
+    name: 'Integration test regimen 2',
+  },
+  {
+    _id: ObjectId('1628394967aa'),
+    name: 'Integration test regimen 3',
+  },
+]