+++ src/backend/resolvers/mutations/sitemap/permissions-upsertion/tests/upsertUserPerms/mocks/input/roles/getRoleTreatmentPlans.js
@@ -1,0 +1,12 @@
+module.exports = (indications) => {
+  return indications.map(({ _id, regimens }) => {
+    const strippedRegimens = regimens.map(({ _id }) => {
+      return { _id }
+    })
+
+    return {
+      _id,
+      regimens: strippedRegimens,
+    }
+  })
+}