+++ src/backend/resolvers/mutations/sitemap/permissions-upsertion/tests/upsertUserPerms/mocks/input/users.js
@@ -1,0 +1,16 @@
+const { CLIENT_A } = require('../../../../../../shared/mocks/clients')
+
+module.exports = {
+  USER_F: {
+    _id: 66,
+    client: CLIENT_A,
+    username: 'User F1',
+    email: 'user_f1@example.com',
+  },
+  USER_G: {
+    _id: 77,
+    client: CLIENT_A,
+    username: 'User G1',
+    email: 'user_g1@example.com',
+  },
+}